Fix(Android): Correct data type for multi-window preferences
This commit fixes a data type mismatch in the `SplitScreenMultiWindow` hook for the Android framework.

The preferences for customizing small window features (`max_simultaneous_small_windows`, `small_window_corner_radius`, `small_window_focused_shadow`, `small_window_unfocused_shadow`) are stored as floats in SharedPreferences but were being read as integers.

This change corrects the preference lookups to use `getFloat()` instead of `getInt()`. The retrieved float value is then cast to an integer before being applied to the corresponding methods, ensuring the hook functions as intended.

+++ app/src/main/java/com/suqi8/oshin/hook/android/SplitScreenMultiWindow.kt
@@ -111,7 +111,7 @@ class SplitScreenMultiWindow : YukiBaseHooker() {
                     }
                 }
             }
-            if (prefs("android\\split_screen_multi_window").getInt("max_simultaneous_small_windows", -1) != -1) {
+            if (prefs("android\\split_screen_multi_window").getFloat("max_simultaneous_small_windows", -1f) != -1f) {
                 "com.android.server.wm.FlexibleWindowManagerService".toClass().apply {
                     method {
                         name = "getMaxWinNum"
@@ -119,12 +119,12 @@ class SplitScreenMultiWindow : YukiBaseHooker() {
                         returnType = IntType
                     }.hook {
                         after {
-                            result = prefs("android\\split_screen_multi_window").getInt("max_simultaneous_small_windows", -1)
+                            result = prefs("android\\split_screen_multi_window").getFloat("max_simultaneous_small_windows", -1f).toInt()
                         }
                     }
                 }
             }
-            if (prefs("android\\split_screen_multi_window").getInt("small_window_corner_radius", -1) != -1) {
+            if (prefs("android\\split_screen_multi_window").getFloat("small_window_corner_radius", -1f) != -1f) {
                 "com.android.server.wm.FlexibleWindowManagerService".toClass().apply {
                     method {
                         name = "getCornerRadius"
@@ -132,12 +132,12 @@ class SplitScreenMultiWindow : YukiBaseHooker() {
                         returnType = IntType
                     }.hook {
                         after {
-                            result = prefs("android\\split_screen_multi_window").getInt("small_window_corner_radius", -1)
+                            result = prefs("android\\split_screen_multi_window").getFloat("small_window_corner_radius", -1f).toInt()
                         }
                     }
                 }
             }
-            if (prefs("android\\split_screen_multi_window").getInt("small_window_focused_shadow", -1) != -1) {
+            if (prefs("android\\split_screen_multi_window").getFloat("small_window_focused_shadow", -1f) != -1f) {
                 "com.android.server.wm.FlexibleWindowManagerService".toClass().apply {
                     method {
                         name = "getShadowRadiusFocused"
@@ -145,12 +145,12 @@ class SplitScreenMultiWindow : YukiBaseHooker() {
                         returnType = IntType
                     }.hook {
                         after {
-                            result = prefs("android\\split_screen_multi_window").getInt("small_window_focused_shadow", -1)
+                            result = prefs("android\\split_screen_multi_window").getFloat("small_window_focused_shadow", -1f).toInt()
                         }
                     }
                 }
             }
-            if (prefs("android\\split_screen_multi_window").getInt("small_window_unfocused_shadow", -1) != -1) {
+            if (prefs("android\\split_screen_multi_window").getFloat("small_window_unfocused_shadow", -1f) != -1f) {
                 "com.android.server.wm.FlexibleWindowManagerService".toClass().apply {
                     method {
                         name = "getShadowRadiusUnfocused"
@@ -158,7 +158,7 @@ class SplitScreenMultiWindow : YukiBaseHooker() {
                         returnType = IntType
                     }.hook {
                         after {
-                            result = prefs("android\\split_screen_multi_window").getInt("small_window_unfocused_shadow", -1)
+                            result = prefs("android\\split_screen_multi_window").getFloat("small_window_unfocused_shadow", -1f).toInt()
                         }
                     }
                 }
